refactor: 重构选择事件监听hook
