+++ src/hooks/select.ts
@@ -3,11 +3,10 @@
  * @version:
  * @Author: June
  * @Date: 2023-04-23 21:10:05
- * @LastEditors: 秦少卫
- * @LastEditTime: 2024-04-10 14:42:34
+ * @LastEditors: wuchenguang1998
+ * @LastEditTime: 2024-05-04 15:37:06
  */
-import Editor, { EventType } from '@kuaitu/core';
-const { SelectEvent, SelectMode } = EventType;
+import Editor from '@kuaitu/core';
 interface Selector {
   mSelectMode: SelectMode;
   mSelectOneType: string | undefined;
@@ -17,54 +16,13 @@ interface Selector {
 }
 
 export default function useSelect() {
-  const state = reactive<Selector>({
-    mSelectMode: SelectMode.EMPTY,
-    mSelectOneType: '',
-    mSelectId: '', // 选择id
-    mSelectIds: [], // 选择id
-    mSelectActive: [],
-  });
-
   const fabric = inject('fabric');
   const canvasEditor = inject('canvasEditor') as Editor;
-
-  const selectOne = (e: [fabric.Object]) => {
-    state.mSelectMode = SelectMode.ONE;
-    if (e[0]) {
-      state.mSelectId = e[0].id;
-      state.mSelectOneType = e[0].type;
-      state.mSelectIds = e.map((item) => item.id);
-    }
-  };
-
-  const selectMulti = (e: fabric.Object[]) => {
-    state.mSelectMode = SelectMode.MULTI;
-    state.mSelectId = '';
-    state.mSelectIds = e.map((item) => item.id);
-  };
-
-  const selectCancel = () => {
-    state.mSelectId = '';
-    state.mSelectIds = [];
-    state.mSelectMode = SelectMode.EMPTY;
-    state.mSelectOneType = '';
-  };
-
-  onMounted(() => {
-    canvasEditor.on(SelectEvent.ONE, selectOne);
-    canvasEditor.on(SelectEvent.MULTI, selectMulti);
-    canvasEditor.on(SelectEvent.CANCEL, selectCancel);
-  });
-
-  onBeforeMount(() => {
-    canvasEditor.off(SelectEvent.ONE, selectOne);
-    canvasEditor.off(SelectEvent.MULTI, selectMulti);
-    canvasEditor.off(SelectEvent.CANCEL, selectCancel);
-  });
+  const mixinState = inject('mixinState') as Selector;
 
   return {
     fabric,
     canvasEditor,
-    mixinState: state,
+    mixinState,
   };
 }
